Move GFT styling to its own initializer
