+++ assets/javascripts/discourse/initializers/global-filter-preference.js
@@ -1,4 +1,4 @@
-import { next, run } from "@ember/runloop";
+import { run } from "@ember/runloop";
 import { ajax } from "discourse/lib/ajax";
 import { popupAjaxError } from "discourse/lib/ajax-error";
 
@@ -27,12 +27,6 @@ export default {
     const globalFilters = siteSettings.global_filters.split("|");
     const currentUser = container.lookup("current-user:main");
     const router = container.lookup("router:main");
-
-    router.on("routeDidChange", () => {
-      next(() =>
-        this._applyFilterStyles(router, globalFilters, currentUser, container)
-      );
-    });
 
     router.on("routeWillChange", (transition) => {
       const routeName = transition.to?.name;
@@ -93,6 +87,7 @@ export default {
           } else {
             filterPref =
               transition.from?.params?.tag_id ||
+              transition.from?.queryParams?.tag ||
               this._firstGlobalFilterFromParent(router, globalFilters) ||
               globalFilters[0];
             this._redirectToFilterPref(
@@ -156,43 +151,6 @@ export default {
 
   _setClientFilterPref(tag, user) {
     return user.set("custom_fields.global_filter_preference", tag);
-  },
-
-  _applyFilterStyles(router, globalFilters, currentUser, container) {
-    // if there is not a filter pref for the current user, tag_id or tag in the params
-    // select the first tag from the parent that matches a global filter
-    let tags;
-    const filterPref = currentUser?.custom_fields?.global_filter_preference;
-    const topic = container.lookup("controller:topic");
-    const topicTags = topic.model?.tags;
-    if (topicTags && topicTags.includes(filterPref)) {
-      tags = filterPref;
-    } else {
-      tags =
-        router.currentRoute?.params?.tag_id ||
-        router.currentRoute?.queryParams?.tag ||
-        this._firstGlobalFilterFromParent(router, globalFilters);
-    }
-
-    if (!tags) {
-      return;
-    }
-
-    globalFilters.forEach((item) => {
-      const filterBodyClass = `global-filter-tag-${item}`;
-      if (item === tags) {
-        document
-          .querySelector(`#global-filter-${item} > a`)
-          .classList.add("active");
-        document.body.classList.add(filterBodyClass);
-        return;
-      }
-
-      document
-        .querySelector(`#global-filter-${item} > a`)
-        .classList.remove("active");
-      document.body.classList.remove(filterBodyClass);
-    });
   },
 
   _firstGlobalFilterFromParent(router, globalFilters) {
